==> notebooks/p2-01-scraping-nba_api.ipynb <==
import pandas as pd
import pickle
import os

from sklearn.preprocessing import MinMaxScaler

import time
import random

from nba_api.stats.static import teams, players
from nba_api.stats.endpoints import playercareerstats, playerawards

import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

# Function to save the dictionary
def save_dataframes(dataframes_dict, filepath):
    with open(filepath, 'wb') as f:
        pickle.dump(dataframes_dict, f)
    # print(f"DataFrames saved to {filepath}")

# Function to load the dictionary
def load_dataframes(filepath):
    if os.path.exists(filepath):
        with open(filepath, 'rb') as f:
            data = pickle.load(f)
        print(f"DataFrames loaded from {filepath}")
        return data
    else:
        print(f"No file found at {filepath}, starting with empty dictionary.")
        return {}

## Fetch a list of all NBA players so their names/IDs can be used to fetch their stats

nba_players = players.get_players()
print("Number of players fetched: {}".format(len(nba_players)))

## Using the list of all players, reduce the list to those who played in the modern era (i.e. since 1980)

cols = playercareerstats.PlayerCareerStats(per_mode36="Totals", player_id=nba_players[2]['id']).get_data_frames()[0].columns

season_to_dataframe = {season: pd.DataFrame(columns=cols) for season in range(1980,2024)}

stats_type = 'PerGame'
df_filepath = f'/Users/cb/src/nba_mvp_ml/data/raw/seasons_{stats_type}'
# save_dataframes(season_to_dataframe, df_filepath)

failed_players = []
succeeded_players = []

## Iterate over the list of players, adding seasons after 1979 to a dict of DataFrames
### Save dataframes to file in case of error

stats_type = 'PerGame'
df_filepath = f'/Users/cb/src/nba_mvp_ml/data/raw/seasons_{stats_type}'

season_to_dataframe = load_dataframes(df_filepath)


# for player in nba_players:
for player in [{'id': 76054,
  'full_name': 'Nate Archibald',
  'first_name': 'Nate',
  'last_name': 'Archibald',
  'is_active': False}]:


    try:
        # Fetch player stats
        if player['id']:
            stats = playercareerstats.PlayerCareerStats(per_mode36=stats_type, player_id=player['id'])
            df = stats.get_data_frames()[0] # 0 indexes main career stats 
        
            # Add the player's name to the df for better human understanding 
            df['PLAYER_FULLNAME'] = player['full_name'].upper()
            columns = df.columns.to_list()
            new_order = columns[:1] + ['PLAYER_FULLNAME'] + columns[1:-1]
            df = df[new_order]
        
            # Conver season column to integers for arithmetic
            df['SEASON_ID'] = [int(year[:4]) for year in df['SEASON_ID']]
            
            # Remove any season before 1980
            # df[df['SEASON_ID'] > 1979]
        
            # Add each player's remaining season to the corresponding season dataframe
            for season in season_to_dataframe.keys():
                season_to_dataframe[season] = pd.concat([season_to_dataframe[season], df[df['SEASON_ID'] == season]], ignore_index=True)
                season_to_dataframe[season] = season_to_dataframe[season][new_order]
                save_dataframes(season_to_dataframe, df_filepath)
    
            succeeded_players.append(player)
            time.sleep(random.randint(0, 3))
            

    except Exception as e:
        print(f'Failed on player {player['id']}: {player['full_name']} with exception {e}.')
        failed_players.append(player)
        


cols = playercareerstats.PlayerCareerStats(per_mode36="Totals", player_id=nba_players[2]['id']).get_data_frames()[0].columns

season_to_dataframe = {season: pd.DataFrame(columns=cols) for season in range(1980,2024)}

stats_type = 'Totals'
df_filepath = f'/Users/cb/src/nba_mvp_ml/data/raw/seasons_{stats_type}'
save_dataframes(season_to_dataframe, df_filepath)

for player in nba_players:

    try:
        # Fetch player stats
        if player['id']:
            stats = playercareerstats.PlayerCareerStats(per_mode36=stats_type, player_id=player['id'])
            df = stats.get_data_frames()[0] # 0 indexes main career stats 
        
            # Add the player's name to the df for better human understanding 
            df['PLAYER_FULLNAME'] = player['full_name'].upper()
            columns = df.columns.to_list()
            new_order = columns[:1] + ['PLAYER_FULLNAME'] + columns[1:-1]
            df = df[new_order]
        
            # Conver season column to integers for arithmetic
            df['SEASON_ID'] = [int(year[:4]) for year in df['SEASON_ID']]
            
            # Remove any season before 1980
            # df[df['SEASON_ID'] > 1979]
        
            # Add each player's remaining season to the corresponding season dataframe
            for season in season_to_dataframe.keys():
                season_to_dataframe[season] = pd.concat([season_to_dataframe[season], df[df['SEASON_ID'] == season]], ignore_index=True)
                season_to_dataframe[season] = season_to_dataframe[season][new_order]
                save_dataframes(season_to_dataframe, df_filepath)
        
            time.sleep(random.randint(0, 3))

        else: 
            print(f'There was no player found for id {player['id']}: ({player['full_name']})')

    except Exception as e:
        print(f'Failed on player {player['id']}: {player['full_name']} with exception {e}.')
        failed_players.append(player)
        


## Test loading the dataframes from file

totals_df_filepath = f'/Users/cb/src/nba_mvp_ml/data/raw/seasons_Totals'
pergame_df_filepath = f'/Users/cb/src/nba_mvp_ml/data/raw/seasons_PerGame'

totals_df = load_dataframes(totals_df_filepath)
totals_df[1980].iloc[:15]

pergame_df = load_dataframes(pergame_df_filepath)

per_game_columns = [
    'MIN', 'FGM', 'FGA', 'FG_PCT', 'FG3M', 'FG3A', 'FG3_PCT', 
    'FTM', 'FTA', 'FT_PCT', 'OREB', 'DREB', 'REB', 'AST', 
    'STL', 'BLK', 'TOV', 'PF', 'PTS'
]

# Rename columns for each DataFrame in the dictionary
for key, df in pergame_df.items():
    df.rename(columns={col: col + '_PG' for col in per_game_columns if col in df.columns}, inplace=True)

pergame_df[1980].iloc[-15:]

merged_dict = {}

for key in totals_df.keys():
    if key in pergame_df:
        # Get corresponding DataFrames
        df1 = totals_df[key]
        df2 = pergame_df[key]
        
        # Perform an inner join on PLAYER_ID to retain only matching rows
        merged_df = pd.merge(df1, df2, on=['PLAYER_ID', 'PLAYER_FULLNAME', 'SEASON_ID', 'LEAGUE_ID', 'TEAM_ID',
       'TEAM_ABBREVIATION', 'PLAYER_AGE', 'GP', 'GS'], how='inner').drop_duplicates(subset='PLAYER_ID', keep='first')
        
        # Add the merged DataFrame to the result dictionary
        merged_dict[key] = merged_df

combined_df = pd.concat(
    [df.assign(Season=season) for season, df in merged_dict.items()],
    ignore_index=True
).drop(columns=['Season'])

df_final_filepath = f'/Users/cb/src/nba_mvp_ml/data/raw/seasons_all'
save_dataframes(merged_dict, df_final_filepath)

output_dir = "/Users/cb/src/nba_mvp_ml/data/processed/combined.csv"
combined_df.to_csv(output_dir, index=False)

from sklearn.preprocessing import MinMaxScaler

columns_to_normalize = ['PLAYER_AGE', 'GP', 'GS', 'MIN', 'FGM', 'FGA',
       'FG_PCT', 'FG3M', 'FG3A', 'FG3_PCT', 'FTM', 'FTA', 'FT_PCT', 'OREB',
       'DREB', 'REB', 'AST', 'STL', 'BLK', 'TOV', 'PF', 'PTS', 'MIN_PG',
       'FGM_PG', 'FGA_PG', 'FG_PCT_PG', 'FG3M_PG', 'FG3A_PG', 'FG3_PCT_PG',
       'FTM_PG', 'FTA_PG', 'FT_PCT_PG', 'OREB_PG', 'DREB_PG', 'REB_PG',
       'AST_PG', 'STL_PG', 'BLK_PG', 'TOV_PG', 'PF_PG', 'PTS_PG']

# Initialize the scaler
scaler = MinMaxScaler()

# Apply normalization only to selected columns
combined_df[columns_to_normalize] = scaler.fit_transform(combined_df[columns_to_normalize])

output_dir = "/Users/cb/src/nba_mvp_ml/data/processed/combined_normalized.csv"
combined_df.to_csv(output_dir, index=False)

1==1
