==> notebooks/p2-03-exploratory-data-analysis-v3.ipynb <==
import sys
import os

# Add the project root directory to sys.path
project_root = os.path.abspath(os.path.join(os.getcwd(), ".."))  # Adjust to your project's structure
sys.path.append(project_root)

from src.analysis import *

player_df, team_df, mvp_df = load_year(2023,debug=True)

list(mvp_df['Player'])

x = per_ws_vs_mvp_points(player_df, team_df, mvp_df)

per_vs_team_success(player_df, team_df, mvp_df)

mvp_df.head()

dims = {
        'PLAYER_FULLNAME': 'Player',
        'eFG%_player': 'Eff. FG%',
        'TS%_player' : 'True Shooting%',
        'BLK': 'Blocks',
        'TRB': 'Rebounds',
        'PTS_opp_pg': 'Opp. PPG',
        'PER': 'PER',
        'WS': 'Win Shares',
        # 'WS/48': 'WS per 48 minutes (WS/48)',
        'W/L%': 'Win-Loss %',
        'Pace': 'Team Pace',
        'MOV': 'MOV',
        'Pts_Won': 'MVP Points Won'
    }

advanced_metrics_player_contribution_filtered(player_df, team_df, mvp_df.head(), dims=dims, include_non_mvp=False)

defense_vs_opponent_scoring(player_df, team_df, mvp_df)

scoring_vs_offensive_rating(player_df, team_df, mvp_df)

top_scorers_team_success(player_df, team_df, mvp_df)

advanced_metrics_scatter_matrix(player_df, team_df, mvp_df)

voting_share_vs_points(mvp_df)

voting_share_vs_ws_per_48(mvp_df)

player_performance_parallel(mvp_df)

age_vs_voting_share(mvp_df)

team_representation(mvp_df)

voting_share_distribution(mvp_df)
