==> notebooks/p3-02-column-cleaning.ipynb <==
import pandas as pd

df = pd.read_csv('/Users/cb/src/nba_mvp_ml/data/processed/by_season/fully_merged/2022-23.csv')

df = df.reindex(sorted(df.columns), axis=1)

# df.to_csv('/Users/cb/src/nba_mvp_ml/data/processed/by_season/fully_merged/SORTED_2022-23.csv')

def find_duplicate_columns(df):
    duplicates = []
    for i in range(len(df.columns)):
        for j in range(i + 1, len(df.columns)):
            if df.iloc[:, i].equals(df.iloc[:, j]):
                duplicates.append((df.columns[i], df.columns[j]))
    return duplicates

# Example Usage
duplicates = find_duplicate_columns(df)
print("Duplicate Columns:", list(duplicates))

list(duplicates)

repeat_cols = [
    '2P%',
    '2P%_opp',
    '3P%_opp',
    '3P%_x',
    '3P%_y',
    'FG3_PCT',
    'AST_player',
    'BLK',
    'FG%_opp',
    'FG%_x',
    'FG%_y',
    'FG_PCT',
    'FT%_opp',
    'FT%_x',
    'FT%_y',
    'FT_PCT',
    'G_y',
    'G_opp',
    'G_pg',
    'G_x',
    'MP_y',
    'MP_x',
    'MP_pg',
    'PA/G',
    'PLAYER_FULLNAME',
    'PS/G',
    'PTS',
    'TRB',
    'TRB_y',
    'SEASON_ID',
    'SRS_wl',
    'STL',
    'TEAM_ABBREVIATION_player',
    'TRB_y',
    'WS_x',
    'WS_y',
    'WS/48_y'
]

print(repeat_cols)
